Profile Api route - wrap with  withAuth HOC

+++ app/api/profile/[uid]/route.ts
@@ -1,38 +1,55 @@
-import { requireUser } from "@/lib/auth/requireUser";
+import { withAuth } from "@/lib/auth/withAuth";
 import prisma from "@/lib/db";
-import { NextRequest, NextResponse } from "next/server";
+import { NextResponse } from "next/server";
 
-export async function DELETE(
-  request: NextRequest,
-  context: { params: Promise<{ uid: string }> }
-) {
-  const { uid } = await context.params;
+type DeleteSuccessResponse = {
+  success: true;
+  message: string;
+};
 
-  const userId = await requireUser(request);
-  if (userId instanceof NextResponse) return userId;
+type DeleteErrorResponse = {
+  success: false;
+  error: {
+    message: string;
+  };
+};
 
-  // check if user exist
-  if (userId !== uid) {
-    return NextResponse.json(
-      { error: "Nemate prava pristupa ovom korisniku" },
-      { status: 401 }
-    );
+export const DELETE = withAuth<DeleteSuccessResponse | DeleteErrorResponse>(
+  async (request, userId) => {
+    const url = new URL(request.url);
+    const uid = url.pathname.split("/").pop();
+
+    if (!uid || uid !== userId) {
+      return NextResponse.json(
+        {
+          success: false,
+          error: { message: "Nemate prava pristupa ovom korisniku" },
+        },
+        { status: 401 }
+      );
+    }
+
+    try {
+      await prisma.user.delete({ where: { uid: userId } });
+
+      return NextResponse.json(
+        {
+          success: true,
+          message: "Profil je uspešno obrisan",
+        },
+        { status: 200 }
+      );
+    } catch (error) {
+      const errorMessage =
+        error instanceof Error ? error.message : "Greška na serveru";
+
+      return NextResponse.json(
+        {
+          success: false,
+          error: { message: errorMessage },
+        },
+        { status: 500 }
+      );
+    }
   }
-
-  try {
-    await prisma.user.delete({
-      where: {
-        uid: userId,
-      },
-    });
-
-    return NextResponse.json(
-      { message: "Profile je uspešno obrisan" },
-      { status: 200 }
-    );
-  } catch (error) {
-    console.log(error);
-    const errorMessage = error instanceof Error && error.message;
-    return NextResponse.json({ error: errorMessage }, { status: 500 });
-  }
-}
+);
